Fix codestyle
