+++ shardingsphere-kernel/shardingsphere-data-pipeline/shardingsphere-data-pipeline-core/src/main/java/org/apache/shardingsphere/data/pipeline/core/api/impl/PipelineResourceAPIImpl.java
@@ -31,7 +31,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 /**
- * Migration source resource API implementation.
+ * Pipeline resource API implementation.
  */
 public final class PipelineResourceAPIImpl implements PipelineResourceAPI {
